requirements.txt
streamlit
pandas
pypdf
python-docx
